refactor(mobile): move location search api and types to shared/location
